Remove `Timestamps` and `Conflicts` from `FpcPayload` (#702)
* remove `Timestamps` and `Conflicts` from `FpcPayload`

* reorganize imports

* address review comments

* address more review comments
